[minor fix] moved code from init file to broker_settings
